print all

+++ 0x10-variadic_functions/3-print_all.c
@@ -79,7 +79,7 @@ void print_int(char *separator, va_list ap)
 
 void print_float(char *separator, va_list ap)
 {
-	printf("%s%f", separator, va_arg(ap, int));
+	printf("%s%f", separator, va_arg(ap, double));
 }
 
 /**
